OpenImageIOReader : Add `cacheMemoryUsage()` static method.
This accompanies the existing `set/getCacheMemoryLimit()` methods, and mirrors the equivalent ValuePlug cache API.

Note that this returns _bytes_, which matches the ValuePlug API, and also the internal OIIO attribute. Unfortunately this is a total mismatch with the set/get methods, which match the internal OIIO attribute ("max_memory_Mb") but not anything else in Gaffer.