Options : Move `prefixPlug()` to CustomOptions node.
It didn't make sense on the base class because it would allow the user to break nodes like StandardOptions, where the correct names are hardcoded already. The approach taken here (virtual methods to define the prefix) could also give us the basis for adding additional subclasses like SceneMetadata (to create "metadata:* entries) in the future. If we do that we would need to rename the Options base class to something more generic.

Breaking Changes :

- Added virtual methods to GafferScene::Options node.