ScriptNode : Allow serialisation and execution from pure C++ code.
Prior to this it was necessary to have a ScriptNodeWrapper instance for serialisation and execution to work, and constructing one of those meant juggling with the boost::python API. ScriptNode instances can now be allocated and used directly in C++ provided that libGafferBindings is linked at runtime.

This doesn't affect the Gaffer applications themselves because they are all implemented in Python, but it makes embedding Gaffer in plugins for other hosts much easier.

Breaking Changes :

- ScriptNode : Removed virtual methods.