Shader : Support disabling of shaders with multiple output parameters.
The key change here is to replace `NetworkBuilder::effectiveShader( shader )` with `NetworkBuilder::effectiveParameter( output )`, since which input a disabled shader routes through depends on which output of the disabled shader is being queried. We now also do the node disabledness testing _before_ calling `Shader::parameterHash()` and `Shader::parameterValue()`, so those methods need no longer be concerned with testing inputs at all - they are guaranteed to be called only with final input parameters which do not have a shader input.

This totally breaks the RenderManShader, because it has its own non-standard handling of coshader inputs (all other renderer backends use a "link:" prefix to denote shader connections, and it doesn't). This special handling is annoying, because it's forcing us to expose the NetworkBuilder to the `parameterHash()` and `parameterValue()` methods, whereas we'd like to entirely hide the internals of the network generation.