ShaderSwitch : Support all parameter types and context-varying indices.
When it was originally written, the ShaderSwitch was designed for use with RSL shaders, where the only connectable parameter was a coshader, represented as a base Plug. Now we support shading languages where any parameter is connectable, we must support all plug types.

Because the ShaderSwitch plugs are not added in the constructor any more, a compatibility config adds them on demand if an old script attempts to access them without first calling `setup()`.