SwitchNodeGadget : Fix python bindings.
This fixes a bug where right clicking on a Switch in the NodeGraph gives the following error :

```
Traceback (most recent call last):
  File "/home/andrewk/apps/gaffer/0.30.2.0dev/cent7.x86_64/cortex/9/gaffer/python/GafferUI/Widget.py", line 903, in eventFilter
    return self.__mouseButtonPress( qObject, qEvent )
  File "/home/andrewk/apps/gaffer/0.30.2.0dev/cent7.x86_64/cortex/9/gaffer/python/GafferUI/Widget.py", line 1003, in __mouseButtonPress
    result = widget._buttonPressSignal( widget, event )
RuntimeError: Exception : Traceback (most recent call last):
  File "/home/andrewk/apps/gaffer/0.30.2.0dev/cent7.x86_64/cortex/9/gaffer/python/Gaffer/WeakMethod.py", line 67, in __call__
    return m( *args, **kwArgs )
  File "/home/andrewk/apps/gaffer/0.30.2.0dev/cent7.x86_64/cortex/9/gaffer/python/GafferUI/NodeGraph.py", line 268, in __buttonPress
    self.nodeContextMenuSignal()( self, nodeGadget.node(), overrideMenuDefinition )
AttributeError: 'GraphComponent' object has no attribute 'node'
```

This was happening because we hadn't exposed SwitchNodeGadget to boost::python, so it was unable to downcast to the appropriate type when returning a SwitchNodeGadget to python. So a call to `gadget.ancestor( NodeGadget )` would find an appropriate NodeGadget on the C++ side, but return it to python as just a GraphComponent due to a lack of information for downcasting.

The problem is, we don't want to expose SwitchNodeGadget to python, because it is a private class. By putting it in a private header and using `copy_class_object()` to bind it, we get to keep it private and appear to Python with a type of StandardNodeGadget (it's first public base class).