GraphGadget : Make nodule/connection handling more robust.
This fixes bugs introduced when refactoring to introduce the NoduleLayout. Because the layout of each edge of a StandardNodeGadget is now handled independently, the GraphGadget must accept the rapid addition and removal of Nodules for the same plug when a nodule is moved from one edge to another. Prior to the refactoring, the existing Nodule was simply reparented to the new edge and the GraphGadget didn't need to be aware of it at all.

The approach used here is to index the internal connection map with nodules rather than plugs, so that we can temporarily track a connection for each nodule when the new nodule is added prior to the old one being removed.

Alternative approaches might have been :

- Refactor NoduleLayout so that it managed multiple sections at once (as we did in the old StandardNodeGadget code), but allowed client code to reparent those sections anywhere they wanted in the Gadget hierarchy.
- Refactor NoduleLayout so that the layout for all instances was performed in a single step, so we could move Nodules from one to another as before.
- Remove ConnectionGadgets entirely, and leave it to Nodules to draw connections.