GafferImage : Add OpenImageIOAlgo
This will allow us to consolidate functionality that is currently scattered privately in GafferImage and GafferOSL. The most important change here is the introduction of DataView, which is similar in concept to a string_view, but providing OIIO compatible views onto IECore::Data. In particular, the view for StringData fixes bugs in our previous code where we assumed that the layout of a std::string was compatible with `const char *` (it no longer is on OSX when using libc++).