Arnold renderer : Fix occasional long hangs in `pause()`
The following appears to have been the problematic sequence :

1. Main thread calls `render()`.
2. Background thread starts up, but doesn't get as far as calling `AiRender()`.
3. Main thread calls `pause()`.
4. Main thread finds `AiRendering() == false` and skips straight to `thread.join()`.
5. Background thread calls `AiRender()` and doesn't return until every refinement level is renderered, effectively hanging the IPR.

This problem could also occur in between refinement levels, where for a short while `AiRendering()` is again not true. We work around this issue by storing our own flag to say if we're rendering or not, and using it as an additional abort flag in the background thread.