TransformTool : Allow for more general updates to the handles
Derived classes may need to edit more than just the transform of the handles, so we need an API that allows for more general updates. The TransformTool base class now also tracks the readonliness of the selection, and triggers calls to `updateTransform()` in the expectation that editability needs to be reflected in the handles somehow.