ErrorDialogue : Fix parentWindow lifetime issues
Because ErrorDialogue.ErrorHandler was deriving from IECore.MessageHandler, it was subject to garbage collection. When used from the FileMenu (and others), the parentWindow passed was a full blown ScriptWindow, which in turn became subject to garbage collection. Garbage collection can kick in at any time, and on any thread. This could lead to a ScriptWindow being destroyed on a non-gui thread, leading to the attempted destruction of QtOpenGL resources on the wrong thread, and therefore deadlocks and crashes.

We fix this by changing the ErrorHandler/MessageHandler relationship from "is a" to "has a", which is arguably what it should have been anyway, even without lifetime considerations.