Main menu : Add options for flushing Arnold render caches
This necessitated the addition of `InteractiveArnoldRender::flushCaches()` to do the actual work while taking care of restarting any running renders to reflect the changes. This is needed because the InteractiveArnoldRender node within the shader swatch viewer is not accessible via the script, so the CacheMenu code could not discover all nodes to perform the restarts itself.