DispatcherUI : Remove emulated "PlaybackRange" option
The Dispatcher class itself doesn't support such an option, because the playback range is something only the UI knows about. We were then trying to emulate the PlaybackRange option by pushing the playback range into the frameRange plug, replacing any custom range the user might have entered. We did keep track of the old custom range so it could be restored when changing back, but we had a bug whereby the playback range could stomp over the user's custom range.

Rather than fix the problem, this commit just removes the PlaybackRange option entirely. It was problematic to implement anyway, and the only reason it had even a chance of working was that the DispatcherWindow was always around in the background to monitor changes to the playback range and push them into the frameFrame plug. In future we intend to allow Dispatcher nodes to exist in the node graph directly, at which point there will be no guarantee of a persistent UI, and the feature would be impossible rather than merely problematic.

I had intended to add a simple preset to the frameRange plug to copy in the current playback range as an alternative, but that is not possible because presets have access only to the plug, not a context. I've also removed the frame range preview feature because it was relying on access to the DispatcherWindow too. Both these should be reintroducable at the expense of further complexity, but personally I don't think it's worth it.