Box : Implement `create()` using `promotePlug()`
We are able to do this because the comment about needing to reparent the node last is no longer relevant, and the mentioned test case passes with the new ordering. This is because in the intervening time 2a40995ecc06249d64b604ceefb4abf582b5e455 has removed the problematic `ShaderAssignment::acceptsInput()` method and replaced it with a custom ShaderPlug plug type which doesn't have any problems with respect to connecting via promoted plugs.