PythonExpressionEngine : Prevent mutation of context
The C++ signature for `Engine::execute()` takes a `const Context *`, but because Python has no const, we use a const_cast so we can push it into Python. The PythonExpressionEngine was then exposing this directly to users as the `context` variable when executing the expression, giving them an opportunity to modify the context, with horrible hard-to-debug knock on effects.

We now provide access to the context through a proxy object which provides access to only non-mutating methods that are tracked by the _Parser class.