ImageStats : Use StringVectorDataPlug for specifying channels
Prior to this we were using a ChannelMaskPlug, and using ImageAlgo::colorIndex() with every matched channel to find a channel whose index matched the output we were computing. This didn't make much sense, because the ChannelMaskPlug could return more than 4 channels, and could end up giving us duplicate channels for a particular color index, which we were silently discarding. We now map the channels[0:4] directly to the RGBA outputs, which has no ambiguity.

Breaking changes :

- Changed interpretation of ImageStats channels to simply be a list of 4 channels that correspond directly to the output RGBA plugs. Behaviour is unchanged apart from situations where an incomplete or ambiguous channel mask was in use.