ScenePlug : Add default constructors and setters to scope classes
This allows us to use the scope classes in cases like `fullTransform()`, where we may not have the path we wish to use until after construction, and we may want to use multiple values for the path.

Having `set` methods without `get` equivalents is a bit of a departure from the norm, but fits with the EditableScope base class and the general idea that the Scope classes are write-only views onto a larger data structure.