OSLShader : Allow vector -> color connections
We were disallowing these on the grounds that OSL doesn't allow it, but that turns out to be incorrect. I don't know if this is something that has changed or I was just wrong in the first place, but the additional ShadingEngine test demonstrates such a connection working, as does the OSLCode node interactively in the shader viewer.