TaskList : Fix `hash()` method
If it really needed to include something in the hash, it should have been calling the base class first and then appending to the result. But because it's a no-op it shouldn't need to append anything anyway. Appending something causes the dispatcher to fire up a Gaffer process just to close it down again.

Fixing this revealed a bug in the Dispatcher, where it refused to coalesce no-op tasks from different frames into the same batch because that would prevent downstream parallelism - see DispatchTest.testNoOpDoesntBreakFrameParallelism. That behaviour is a good one, but not when the no-op requires sequence execution, since that implies that the tasks _must_ be coalesced into the same batch.