UIEditor : Fix FileSystemPathPlugValueWidget extensions bug
At some point we made the unwise decision of allowing the extensions to either be StringVectorData or a space separated string. The UIEditor wants to edit strings since that is more convenient for the user, so we must cast the StringVectorData to string if necessary.

This bug could be exposed as follows :

1. Create Box containing a SceneReader
2. Promote the SceneReader.fileName plug
3. Open the UIEditor, go to the Plugs section, and select the fileName plug