Dispatcher : Improve batching of no-ops such as TaskList
We were already forcing these into a unique batch per frame, because otherwise an intermediate no-op would break the frame-by-frame parallelism of the graph, forcing all frames of the upstream task to complete before any downstream task could start. But this didn't address a similar problem when an intermediate no-op was upstream of a wedge - batches were coalesced across wedge variables, yielding a dispatch order that while technically correct, was not the preferred one.

This commit fixes the coalescing across wedge variables, and simplifies the previous special case so that it's now the standard batch size limiting that prevents frames from coalescing. This has the nice side effect of making the `dispatcher.batchSize` plug relevant for TaskList and other no-op nodes.