GLWidget : Update HostedGLContext for Qt 5
In Qt 5 all the old QGL classes are now deprecated and there are new QOpenGL classes in their place. We want to continue to use the QGL ones so that we can support Qt 4 and Qt 5 in parallel until all Gaffer hosts in production use are using Qt 5. In Qt 5 the QGLContext class is now implemented behind the scenes using a QOpenGLContext class, which causes problems for our old "share the context with the host" code :

- QGLContext::chooseVisual() has been removed, so we must manually choose an appropriate GLXFBConfig instead
- QGLContext expects to be associated with a QOpenGLContext privately behind the scenes, and internal components expect this association to be valid. As far as I can tell, this breaks the concept of allowing the `QGLContext::create()` and `QGLContext::makeCurrent()` virtual methods to manage the actual underlying context themselves, because they cannot keep the QOpenGLContext in sync due to it being private. Fortunately it seems sufficient to call the base class `makeCurrent()` method to make the private QOpenGLContext current before we then override it with our own context.