GafferUI.__init__ : Use Qt.py
This provides access to PySide2, PySide, PyQt5 or PyQt4 through a uniform interface matching PySide2, and including only those members which are available in all bindings. It is therefore well suited to the task of migrating to PySide2 while maintaining compatibility with PySide and PyQt4.

See https://github.com/mottosso/Qt.py for more details.