Display : Move update mechanism to C++
This shaves nearly 20% off the rendertime of a lightweight interactive render with 20+ aovs, and increases the viewer framerate for the same render from 11-12fps to 40-45fps. A nice side effect is that we now emit Display::driverCreatedSignal, Display::dataReceivedSignal and Display::imageReceivedSignal on the UI thread, so clients don't need to jump through hoops to use them.