RenderManRenderTest : Mark expected failure on Mac
This only affects a platform we don't use in production, using a hider that was long since unused in production even on unaffected platforms, and in a renderer backend we intend to replace shortly anyway. If anyone is affected and wants to provide a fix it would be gratefully accepted, but we don't have the resources to address this ourselves.