Catalogue : Reuse cache entries after saving
This avoids the sometimes significant delay introduced by immediately loading the same image from disk again.

As an alternative strategy, I did try precaching, whereby the image was loaded into Gaffer's cache on the background thread before switching to read the file on the UI thread. This was much less effective, and also resulted in spending about twice the time on the background thread as before.