Catalogue : Fix AsynchronousSaver shutdown crashes
We weren't waiting for the background thread to complete when shutting down, meaning it was merrily trying to compute a graph while the entire framework it needed was being torn down around it on the main thread. We now ensure that the destruction of a Catalogue (which occurs on the main thread during shutdown) waits for the background thread to complete before returning.