Display : Replace boost::unique_ptr with boost::shared_ptr
We want to use std::unique_ptr, but we can't because we have to maintain compatibility with C++98 for a bit longer. We can't use std::auto_ptr, because it's deprecated in GCC 5. And we can't use boost::movelib::unique_ptr, because that was only added in boost 1.57 and we have to maintain compatibility with VFXPlatform 2016 for a bit longer, and that mandates boost 1.55 (we get 1.61 in VFXPlatform 2017).