OpenImageIOReader : Add fileFormat and dataType metadata
Also updated ImageWriter to ensure that these are not written into files since they could contradict the file contents when writing to a different format, and updated tests as necessary.

Question for reviewers - should these metadata names be prefixed with something? Standard oiio metadata is already prefixed with "oiio:", so perhaps that makes sense given that this is derived directly from information provided by OIIO? Or perhaps "openImageIOReader:" because it's the reader that has created them?