ContextVariablesUIs : Document extraVariables plug
Unfortunately we have to do this in triplicate for ContextVariables/SceneContextVariables/ImageContextVariables, because the templated approach to defining the nodes means there is no common base class. Issue #2123 suggests we should phase out the dedicated scene/image nodes and instead just expose the generic node to the user directly, at which point this would no longer be a problem.