GafferScene bindings : Reorganise for simplicity and compilation speed
Previously the bindings were in a separate libGafferSceneBindings library, complete with headers that we installed with Gaffer, as if the library might be useful to others. In actual fact the library was useless outside of the GafferScene module, so this commit moves all the code into the src/GafferSceneModule directory so that it is entirely private.

At the same time, the binding code is now consolidated into fewer files to improve compilation speed. Code using boost::python is really slow to compile, so we've been splitting the bindings up into many files to minimise the compilation speed of any given file. This does give faster recompilation when only one binding file is affected by a change, but is a false economy for full builds, as the following table testing different numbers of node bindings per file shows :

|Nodes bound in file   | Compilation time  |  Time per node |
|----------------------|-------------------|----------------|
|1                     | 3.94              | 3.94           |
|2                     | 4.61              | 2.31           |
|3                     | 5.05              | 1.68           |
|4                     | 5.50              | 1.375          |
|5                     | 5.95              | 1.18           |
|6                     | 6.41              | 1.06           |

Basically there's a large up-front cost to just getting all the boost headers included and parsed, so we can get better overall compilation times by amortising that across multiple bindings in a single file. Peak memory usage does go up as we add bindings, and it is annoying to have too many nodes in a single file when doing incremental recompilation, so I've aimed for a bit of a sweet spot rather than forced all bindings into one file. With this change total compilation time for the GafferScene module is almost halved.