Travis : Use SCons cache
This combines SCons' ability to use a build cache of previously compiled objects, and Travis' ability to squirrel away caches and reuse them for future builds. The Travis caches are separate per platform/compiler/environment so each row of our build matrix has a separate cache. This has a huge positive impact on our build times when only a few files have changed on a branch.