DisplayTest : Improve UI emulation
To be able to run the tests we need to emulate the way in which DisplayUI hooks into Display::executeOnUIThreadSignal() to perform actions on the main thread when they are requested from the background server thread. Until now we were playing fast and loose with this and were just executing directly on the background thread. We'd got away with this for a long time but the addition of the Catalogue tests brought many intermittent problems on Travis that have not been seen to occur in the UI itself.