TaskNode : Derive from DependencyNode
This allows us to declare dependencies between the input plugs and the output task plug. The primary use for this at this stage is to allow the NodeGraph to clear the error indicator after a setting has been adjusted following an unsuccessful dispatch.

This also opens the door to adding an "enabled" plug, but we need to decide on appropriate semantics for disabled TaskNodes before we take advantage of that.