Shape : Fix crashes caused by invalid assertions
The compute* and hash* methods were actually being used to compute the shadow shape as well as the main shape. In fact, the only thing different about the shadow shape is the colour, so we can share cache entries by driving the shadow shape with a connection from the main shape, and then fix the assertion to allow computing the shadow shape in the channelData methods.

Thanks to @est77 for reporting this problem.