ConnectionCreator : add abstraction for classes that can add connections
This is meant to provide a clean interface for setting up connections.

We consider this useful for UI interaction where a Gadget that received
a drop event just wants the sourceGadget to connect to itself without
knowing what the source is. That source might want to do some setup and
cleanup that the destination doesn't even know about and via this
interface that is neatly hidden.