StandardNodeGadget : Don't abuse drag 'n' drop signals
These are expected to be emitted by the core event systems in Widget and ViewportGadget, where it is assumed that `event.destinationGadget` is the last gadget to have accepted `dragEnterSignal()`. Instead we simply use highlighting to indicate the drag destination proxy.

This has an unfortunate downside though : we no longer get the labels being shown for the nodules. We would need to fix that before continuing with this approach.

This commit is also incomplete in that it isn't dealing with tangents properly.