StandardNodeGadget : Don't abuse drag 'n' drop signals
These are expected to be emitted by the core event systems in Widget and ViewportGadget, where it is assumed that `event.destinationGadget` is the last gadget to have accepted `dragEnterSignal()`. We got away with cheating this system for a long time, but it has always been fragile. Fortunately the new ConnectionCreator API provides us with what we need to do this more cleanly.