ConnectionCreators : Rationalise use of UndoScope
It makes most sense to construct the UndoScope at the outermost scope just before `createConnection()` is called - as with elsewhere in Gaffer the caller is then in control of what happens in terms of undo. Instead we had a mishmash where some callers were not creating an UndoScope (a bug), but the bug was masked by several implementations creating an UndoScope internally anyway.