GafferImageTest : Avoid IECore::ImagePrimitive where possible
This is mostly a case of replacing comparisons of ImagePrimitives with calls to `assertImagesEqual()` or `assertImageHashes()` equal. There are still some places where we don't have a direct replacement, so those remain, but they are far fewer now. This reflects the fact that ImagePrimitive is now taking even more of a back seat role, with GafferImage/OIIO being the preferred interfaces.