SConstruct : Bump version to 0.40.0.0
This will be the first released version to require C++11 and Cortex 10. The astute observer will note that this leaves a gap from the previous major version of 0.34 : this is to provide growing space for future C++03/Cortex9 versions in case they may be required by productions which cannot yet make the jump to 0.40.