ImageReader : Output blank image when disabled
The overriding of the bare `hash()` and `compute()` methods was hiding the implementation of disabled output implemented in the ImageNode base class, so we must implement the plug-specific methods instead. In addition to fixing the bug, this also avoids the unnecessary work we were doing in the hash function before, where we were calling the base class to start building a hash, and then overwriting it anyway in `hashMaskedOutput()`.

Fixes #2056