ArnoldRenderer : Prefix names for nodes made by procedurals
This is necessary to avoid clashes with names generated outside procedurals.

We could simply pass a name prefix to the various node generators, but we're passing a parent node instead. This is because in Arnold 5 we will need to create nodes beneath the procedural parent instead, and won't need to manually prefix the names, because they will be namespaced by the procedural.