OSL ShadingEngine : Accept UV globals as V2fVectorData
This is needed to correctly fill in the globals when shading MeshPrimitives in OSLObject. The alternative would be to split the UVs into separate FloatVectorDatas in OSLObject but this represents unnecessary overhead. Question : since OSLImage has to generate the UV data on the fly anyway, should we just simplify things and _only_ accept V2fData?