ValuePlug : Emit `plugSet()` for all ancestors, not just ValuePlugs
This is consistent with the documentation for `ValuePlug::emitPlugSet()` and the signature for `Node::plugSetSignal()`.

The immediate motivation for this change is to fix test failures in GafferCortexTest.ParameterisedHolderTest when we move to using Plug rather than CompoundPlug to represent CompoundParameters. Without it, `parameterChanged()` is not called for compound parameters.