Outputs : Remove backwards compatibility
The main reason for this is that prior to 553407a5cd50da833c363fc3529d08dd5f46e66d (August 2015) we were using CompoundPlugs rather than ValuePlugs to represent outputs, and we are now finally removing CompoundPlugs entirely. If we wanted we could use a `Gaffer.CompoundPlug = Gaffer.ValuePlug` compatibility workaround, but I think the changeover is sufficiently far in the past that this isn't necessary.

Note that we're also dropping compatibility with some other changes here (plug naming and Display->Outputs node rename), because they preceded the CompoundPlug->ValuePlug change anyway.