OSLShaderUI : Hide spline parameters in NodeGraph
There's nothing sensible to connect them to. This also fixes the following error when trying to promote a spline parameter :

```
RuntimeError: Exception : Cannot set value for plug "out.interpolation" except during computation.
```

The problem here was that when promoting via a BoxIn node (the code path if there's a nodule for the parameter), an output SplinePlug was created, and in the constructor `setValue()` was called.