Expression : Fix copy/paste bug
The problem here was that we were including a `setExpression()` call in the serialisation, which we were using to restore m_engine. But the expression contained the relative names of the external plugs to connect to, which would be renamed during pasting to avoid conflicts with the original plugs. This would mean the new node connected to the original plugs and not the copy.