BoxIO : Fix undo/redo of `insert()` method
We need to make sure that the nodes have a parent when we perform actions we need to be replayed by the undo system. But we also need to make sure they _don't_ have a parent when we call `setup()`, because that would cause the creation of duplicate external plugs.