PlugAlgo : Fix promotion of non-serialisable output plugs
It's no use making a plug dynamic if the serialisable flag is off.

Bugs around the dynamic/serialisable flags are pretty common, and forgetting to set the dynamic flag is a common mistake for folks new to scripting. I'm beginning to think we'd be much better off if we could just abolish the flags entirely, and instill the right behaviour using the Serialiser classes, using knowledge of what a plug is parented to.