GraphComponent binding : Fix crashes caused by `None` arguments
In all these functions we want either a string or InternedString argument, but had bound them with wrappers taking `const char *`. But boost python will happily convert `None` to null and pass it to such methods, which resulted in crashes when subsequently passing null to the string or InternedString constructors.