Add "how to run" example for linux and link to installation page
Running the downloaded binary version is a very straight forward process, for an experimented user. Others can need minute before they will be sure sure what are the sequence of commands to follow (for example, is it gaffer or gaffer.py the starting point?)

Also added the link to the installation documentation, where useful information can be found. Like which render engine is available and how I can connect my preferred one.