NodeGraph : Avoid confusing box navigation
Previously we would happily dive down any box, even if it was totally unrelated to the current root, or several nesting levels deep. We now only jump if the new root is a descendant of the current root, and if it is nested several levels deep we step towards it one level at a time.

Fixes #2343