SceneView : Refactor gnomon
The gnomon is a bit of a corner case because it does all sorts of funky stuff with the projection and model view matrices. It assumed that its child gadgets would inherit those transforms, but since that is not going to be the case for much longer, we must adjust.