ValuePlug : Improve serialisation of compound plugs
We now reliably emit a `setValue()` call at the highest point possible in the hierarchy, as dictated by plug inputs and other factors. This fixes a couple of specific issues :

- BoxPlugs were serialised with `setValue()` calls for the leaf plugs _and_ the top level plug itself.
- FormatPlugs were incorrectly attempting to serialise a `getValue()` call at the top level even when input connections existed at the leaf level. This prevented a common pattern for the dispatch of image networks from working.

The `ValuePlugSerialiser::valueNeedsSerialisation()` method was a previous attempt to fix this problem, but it only applied to CompoundNumericPlugs. It would probably have been possible to fix the problem using this mechanism, but the approach is flawed. The decision about whether or not to serialise at a particular level of the plug hierarchy can only be made with reference to all descendants and ancestors, so to avoid unnecessary work we must tackle the problem with a single traversal under our direct control.

Fixes #1500

Breaking Changes :

- Removed `ValuePlugSerialiser::valueNeedsSerialisation()` virtual method.