TypedObjectPlug : Use `extern template`
Even though we're deliberately hiding all the implementation of the class in .cpp files, this is not enough to stop the compiler automating generating typeinfo symbols in every translation unit that includes the TypedObjectPlug header. This meant that libGaffer had one copy of the symbol (the one we wanted), but python/Gaffer/_Gaffer.so had its own copy. On Mac this broke the python bindings with this error :

```
ArgumentError: Python argument types in
    InternedStringVectorDataPlug.defaultValue(InternedStringVectorDataPlug)
did not match C++ signature:
    defaultValue(boost::intrusive_ptr<Gaffer::TypedObjectPlug<IECore::TypedData<std::__1::vector<IECore::InternedString, std::__1::allocator<IECore::InternedString> > > > >, bool _copy=True)
```

Because we had two typeinfo symbols for the same class, type comparisons were failing and boost::python believed it was not receiving an object of the right type. This didn't occur on Linux, because there boost python compares typeinfo names rather than typeinfo addresses.

The solution is to declare the TypedObjectPlug classes as `extern template`, preventing the compiler from emitting typeinfo symbols except when compiling libGaffer.