VDBObject : Improve metadata conversion
Based on the canonical example at http://www.openvdb.org/documentation/doxygen/codeExamples.html#sGettingMetadata :

- Use `staticTypeName()` rather than hardcoded strings
- Use `static_cast()` rather than `DynamicPtrCast()` (because we've already checked the type)

Other than simplifying things, the motivator here is that the current code was silently failing on OSX. This was because the dynamic_cast was failing, because openvdb doesn't mark the TypedMetadata classes as exported. I presume this led to libGafferVDB.so
and libopenvdb.so having different copies of the typeinfo. Typeinfo comparisons on OSX compare the symbol address, so we would need both libraries to share the same symbol for the comparison to succeed. We get away with it on Linux because there typeinfo comparisons compare the typename strings instead.

Note that this was also fixable by enclosing the openvdb headers in IECORE_PUSH/POP_DEFAULT_VISIBILITY, but it seemed worth following the canonical example instead since it also simplifies the code.